@@ -1,9 +1,7 @@
 /** @type {import('jest').Config} */
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  roots: ['<rootDir>/src'],
-  testMatch: ['**/*.test.ts'],
-  moduleFileExtensions: ['ts', 'js', 'json'],
-  collectCoverageFrom: ['src/**/*.ts', '!src/test/**', '!src/**/*.test.ts'],
-};
+export const preset = 'ts-jest';
+export const testEnvironment = 'node';
+export const roots = ['<rootDir>/src'];
+export const testMatch = ['**/*.test.ts'];
+export const moduleFileExtensions = ['ts', 'js', 'json'];
+export const collectCoverageFrom = ['src/**/*.ts', '!src/test/**', '!src/**/*.test.ts'];
